+++ statistic.py
